refactor(main): extract desktop tool-assembly cluster (arch R4)
Pure move of main.ts's module-scope tool-assembly block into
tool-assembly.ts: assembleDesktopTools(deps) builds the sandbox /
filesystem worker, the deferred capability groups (Rive, Office, browser,
computer-use, agent orchestration), the WebSearch tool, the builtin + skill
host surface, the deferred-group toolAvailability, and childAgentTools, and
returns the 11 collections main.ts consumes downstream. main.ts keeps every
call site plus the seams that must stay module-scoped: the backends.register
ai-sdk closure (modelSupportsVision unchanged), systemPromptService,
storeReadImage, and the onMainWindowClose cursor-overlay teardown (assigns a
module let). Declaration order inside assembleDesktopTools preserves the
original init order. main.ts 1903 -> 1656 (-247). Zero behavior change.
