+++ apps/desktop/src/main/tool-assembly.ts
@@ -1,0 +1,277 @@
+import { app, nativeImage, powerMonitor } from 'electron';
+import {
+  buildAgentTeamChildTools,
+  buildAgentTeamLeadTools,
+  buildAskUserQuestionTool,
+  buildBuiltinTools,
+  buildChildAgentTools,
+  buildDeferredToolGroupsFromCatalog,
+  buildHostCapabilitiesFromBinding,
+  buildParentAgentTools,
+  assertProductBindingCatalogClean,
+  createBuiltinSandboxManager,
+  createFilesystemWorkerLaunchSpecProvider,
+  FilesystemWorkerClient,
+  resolveSkillDiscoveryPaths,
+  ShellRunProcessManager,
+  SKILL_TOOL_NAME,
+} from '@maka/runtime';
+import type {
+  HostCapabilitiesResolver,
+  MakaTool,
+  ToolAvailabilityConfig,
+} from '@maka/runtime';
+import type { WorkspacePrivacyContext } from '@maka/core/incognito';
+import { createAgentMailboxStore, createSettingsStore } from '@maka/storage';
+import { createComputerUseOverlayHook } from '@maka/computer-use';
+import { buildWebSearchAgentTool } from './web-search/agent-tool.js';
+import { buildRiveWorkflowTool } from './rive-workflow-tool.js';
+import { buildOfficeDocumentEditTool, buildOfficeDocumentTool } from './office-document-tool.js';
+import { buildExploreAgentTool } from './explore-agent-tool.js';
+import { buildBrowserTools } from './browser/browser-tools.js';
+import { createComputerUseHost } from './computer-use-host.js';
+import { createCursorOverlayController } from './computer-use/cursor-overlay-window.js';
+import {
+  applyComputerUseRealModelPolicy,
+  parseComputerUseRealModelPolicy,
+} from './computer-use-real-model-policy.js';
+import { buildSkillAgentTool } from './skills.js';
+import type { createMainTaskLedgerWiring } from './task-ledger-wiring.js';
+import type { createMainAutomationWiring } from './automation-wiring.js';
+import type { createMainGoalWiring } from './goal-wiring.js';
+import type { ToolArtifactPersistence } from './tool-artifact-persistence.js';
+
+type TaskLedgerWiring = ReturnType<typeof createMainTaskLedgerWiring>;
+type AutomationWiring = ReturnType<typeof createMainAutomationWiring>;
+type GoalWiring = ReturnType<typeof createMainGoalWiring>;
+type AgentMailboxStore = ReturnType<typeof createAgentMailboxStore>;
+type SettingsStore = ReturnType<typeof createSettingsStore>;
+
+export interface DesktopToolAssemblyDeps {
+  /** E2E computer-use flag: routes the ai-sdk backend through the raw
+   *  computer-use tools and disables the economy, matching the legacy path. */
+  isComputerUseRealModelE2e: boolean;
+  workspaceRoot: string;
+  taskLedgerStore: TaskLedgerWiring['store'];
+  taskLedgerWiring: TaskLedgerWiring;
+  automationWiring: AutomationWiring;
+  goalWiring: GoalWiring;
+  agentMailboxStore: AgentMailboxStore;
+  settingsStore: SettingsStore;
+  shellRuns: ShellRunProcessManager;
+  snapshotReadImage: ToolArtifactPersistence['snapshotReadImage'];
+  getWorkspacePrivacyContext: () => Promise<WorkspacePrivacyContext>;
+  resolveDesktopSkillHost: HostCapabilitiesResolver;
+}
+
+/**
+ * Assemble the desktop process's tool surface (issue #37 economy split).
+ * Pure move of main.ts's module-scope tool-assembly cluster: the sandbox /
+ * filesystem worker, the deferred capability groups (Rive, Office, browser,
+ * computer-use, agent orchestration), the WebSearch tool, the builtin + skill
+ * host surface, the deferred-group tool-availability config, and the child
+ * agent tool surface. Declaration order inside the function preserves the
+ * original module-init order. The cursor-overlay `onMainWindowClose` teardown
+ * hook stays in main.ts (it assigns a module-scoped `let`); the overlay
+ * controller is returned so main.ts can wire it.
+ */
+export function assembleDesktopTools(deps: DesktopToolAssemblyDeps) {
+  const {
+    isComputerUseRealModelE2e,
+    workspaceRoot,
+    taskLedgerStore,
+    taskLedgerWiring,
+    automationWiring,
+    goalWiring,
+    agentMailboxStore,
+    settingsStore,
+    shellRuns,
+    snapshotReadImage,
+    getWorkspacePrivacyContext,
+    resolveDesktopSkillHost,
+  } = deps;
+
+  const sandboxManager = createBuiltinSandboxManager();
+  const filesystemWorker = process.platform === 'darwin' && sandboxManager
+    ? new FilesystemWorkerClient({
+        sandboxManager,
+        getLaunchSpec: createFilesystemWorkerLaunchSpecProvider({
+          runtime: 'electron',
+          executable: process.execPath,
+          resourceLocation: app.isPackaged
+            ? { kind: 'desktop-packaged', resourcesPath: process.resourcesPath }
+            : { kind: 'runtime' },
+        }),
+      })
+    : undefined;
+  // Unified tool availability (issue #37). Deferred capability groups (Rive,
+  // Office, browser, agent orchestration) are withheld from the
+  // per-turn prompt and loaded on demand via `load_tools`, keeping their schemas
+  // off the wire until needed. Everything else (ungrouped) stays always-on.
+  // Kill-switch: set MAKA_DISABLE_DEFERRED_TOOLS to any value to turn economy off
+  // and advertise every tool every turn (legacy behavior).
+  const economyEnabled = !process.env.MAKA_DISABLE_DEFERRED_TOOLS;
+  const riveTools: MakaTool[] = [buildRiveWorkflowTool()];
+  const officeTools: MakaTool[] = [buildOfficeDocumentTool(), buildOfficeDocumentEditTool()];
+  // Embedded-browser observe→act tools. They drive the conversation's own
+  // WebContentsView via the BrowserViewHost the desktop provides in registerIpc;
+  // outside the app (no host) they report the browser as unavailable.
+  const browserTools: MakaTool[] = buildBrowserTools();
+  const computerUseOverlay = createCursorOverlayController();
+  const computerUseHost = createComputerUseHost({
+    isPackaged: app.isPackaged,
+    resourcesPath: process.resourcesPath,
+    compressFrame: (base64) => {
+      try {
+        const image = nativeImage.createFromBuffer(Buffer.from(base64, 'base64'));
+        return image.isEmpty()
+          ? { base64, mimeType: 'image/png' }
+          : {
+              base64: image.toJPEG(82).toString('base64'),
+              mimeType: 'image/jpeg',
+            };
+      } catch {
+        return { base64, mimeType: 'image/png' };
+      }
+    },
+    physicalInputRecentlyActive: () => powerMonitor.getSystemIdleTime() < 1,
+    ...(isComputerUseRealModelE2e
+      ? {
+          onTrace: (event) => {
+            const tracePath = process.env.MAKA_CU_REAL_MODEL_TRACE;
+            if (!tracePath) return;
+            void import('node:fs/promises').then(({ appendFile }) =>
+              appendFile(tracePath, `${JSON.stringify(event)}\n`, {
+                encoding: 'utf8',
+                mode: 0o600,
+              }),
+            ).catch(() => {});
+          },
+        }
+      : {}),
+    overlay: createComputerUseOverlayHook(computerUseOverlay),
+  });
+  const computerUse = computerUseHost.selected;
+  const computerUseTools = applyComputerUseRealModelPolicy(
+    computerUse.tools,
+    isComputerUseRealModelE2e
+      ? parseComputerUseRealModelPolicy(
+          process.env.MAKA_CU_REAL_MODEL_POLICY,
+        )
+      : undefined,
+  );
+  const agentTools: MakaTool[] = buildParentAgentTools({
+    taskLedger: taskLedgerStore,
+  });
+  const agentTeamLeadTools = buildAgentTeamLeadTools({
+    mailbox: agentMailboxStore,
+    taskLedger: taskLedgerStore,
+  });
+  const agentTeamChildTools = buildAgentTeamChildTools({
+    mailbox: agentMailboxStore,
+    taskLedger: taskLedgerStore,
+  });
+  const deferredTools: MakaTool[] = [
+    ...riveTools,
+    ...officeTools,
+    ...browserTools,
+    ...computerUseTools,
+    ...agentTools,
+  ];
+  const webSearchTool = buildWebSearchAgentTool({
+    settingsStore,
+    getPrivacyContext: getWorkspacePrivacyContext,
+  });
+  // Assemble product tools first, then derive skill host + deferred groups from
+  // the shared catalog ∩ this binding (#1099 S2). Skill listing uses the same host.
+  const toolsBeforeSkill: MakaTool[] = [
+    buildAskUserQuestionTool(),
+    ...buildBuiltinTools({
+      shellRuns,
+      runtimeResources: shellRuns,
+      backgroundTasks: shellRuns,
+      ptyControls: shellRuns,
+      snapshotImage: snapshotReadImage,
+      ...(sandboxManager ? { sandboxManager } : {}),
+      ...(filesystemWorker ? {
+        filesystemWorker,
+        enableBashAdditionalPermissions: true,
+        enableFileToolAdditionalPermissions: true,
+      } : {}),
+    }).filter((tool: MakaTool) => tool.name !== 'Edit'),
+  ];
+  const toolsAfterSkill: MakaTool[] = [
+    // External reference plan-mode borrow: a bounded read-only local worker for
+    // self-contained code/repo investigations. The tool advertises the
+    // `subagent` category; explore mode allows it, but the implementation
+    // itself only reads filenames/text snippets under the session cwd.
+    buildExploreAgentTool(),
+    // PR-AGENT-WEB-SEARCH-TOOL-0: Tavily-backed WebSearch tool. Closed
+    // over settingsStore so the renderer never sees the API key; the
+    // permission engine routes it through the `web_read` policy which
+    // prompts the user in explore / ask modes.
+    webSearchTool,
+    // Session task ledger: model manages a flat task list; the current list is
+    // re-injected each turn tail. Pure local state, so no permission gate.
+    ...taskLedgerWiring.tools,
+    // Unified Automation: heartbeat (session-internal polling) + cron (standalone scheduled runs).
+    ...automationWiring.tools,
+    // Goal execution: GoalSet/Clear/Status/Pause/Resume — autonomous turn-boundary continuation.
+    ...goalWiring.tools,
+    // The `load_tools` connector is built by ToolAvailabilityRuntime; deferred
+    // group tools just need to be present so they are dispatchable once loaded.
+    ...deferredTools,
+  ];
+  // Always-on Skill name is part of the host surface even before the tool instance
+  // is built (so requiredTools gates and capability tags stay complete).
+  const desktopBoundToolNames = [
+    ...toolsBeforeSkill.map((tool) => tool.name),
+    SKILL_TOOL_NAME,
+    ...toolsAfterSkill.map((tool) => tool.name),
+  ];
+  assertProductBindingCatalogClean('desktop', desktopBoundToolNames);
+  const desktopHostCapabilities = buildHostCapabilitiesFromBinding(desktopBoundToolNames);
+  // External reference lazy-skill pattern: the prompt lists available skills,
+  // and this read-only tool loads the full SKILL.md only when the task matches.
+  // Resolve per-call from the session cwd so skills at all 5 standard paths
+  // (cwd/.maka, cwd/.agents, workspaceRoot/skills, ~/.maka, ~/.agents) are
+  // discovered — matching the CLI and the Agent Skills spec (#1068).
+  const skillTool = buildSkillAgentTool(
+    ({ cwd }) => resolveSkillDiscoveryPaths(cwd, workspaceRoot),
+    resolveDesktopSkillHost,
+  );
+  const builtinTools: MakaTool[] = [...toolsBeforeSkill, skillTool, ...toolsAfterSkill];
+  const toolAvailability: ToolAvailabilityConfig = {
+    economy: economyEnabled,
+    groups: buildDeferredToolGroupsFromCatalog('desktop', desktopBoundToolNames),
+  };
+  // Child agents stay file-only for local reads; parent runtime refs such as
+  // maka://runtime/background-tasks/<id> are not part of their tool surface.
+  const childAgentTools = buildChildAgentTools([
+    ...buildBuiltinTools({
+      snapshotImage: snapshotReadImage,
+      ...(sandboxManager ? { sandboxManager } : {}),
+      ...(filesystemWorker ? {
+        filesystemWorker,
+        enableBashAdditionalPermissions: true,
+        enableFileToolAdditionalPermissions: true,
+      } : {}),
+    }).filter((tool: MakaTool) => tool.name !== 'Edit'),
+    webSearchTool,
+    ...agentTeamChildTools,
+  ]);
+
+  return {
+    riveTools,
+    officeTools,
+    browserTools,
+    computerUse,
+    computerUseOverlay,
+    computerUseTools,
+    agentTeamLeadTools,
+    desktopHostCapabilities,
+    builtinTools,
+    toolAvailability,
+    childAgentTools,
+  };
+}
